requirements.txt
# Core dependencies for Omotenashi prototype v0.1
# AI and workflow orchestration
anthropic>=0.18.0          # For Claude 3.5 Sonnet API
langchain>=0.1.0           # For ReAct agent and tools
langchain-anthropic>=0.1.0 # For Claude integration with LangChain

# CLI and configuration
click>=8.1.0               # For command-line interface
pyyaml>=6.0                # For YAML configuration files
python-dotenv>=1.0.0       # For environment variable management

# Data handling
pydantic>=2.0.0            # For data validation and models
typing-extensions>=4.0.0   # For enhanced type hints

# Development utilities
colorama>=0.4.6            # For colored CLI output

# Telegram bot integration
python-telegram-bot>=20.7  # For Telegram bot functionality
aiosqlite>=0.19.0          # For async SQLite operations